fix: change lastScrollPosition type to Float for improved precision in scroll tracking

+++ app/src/main/java/io/aatricks/novelscraper/data/model/LibraryItem.kt
@@ -28,7 +28,7 @@ data class LibraryItem(
     val dateAdded: Long = System.currentTimeMillis(),
     val lastRead: Long = System.currentTimeMillis(),
     val isDownloading: Boolean = false,
-    val lastScrollPosition: Int = 0,
+    val lastScrollPosition: Float = 0f,
     val lastReadIndex: Int = 0,
     val lastReadOffset: Int = 0,
     val hasUpdates: Boolean = false,
